feat: Implement flexible paths
Allow the possibility to choose how many points of the last planned path should be taken over for the new path, instead of always taking all previous planned path points over to the new path. This allows for more flexible path planning that can react better to sudden changes in the environment.

This is currently not working flawlessly, check the TODOs in behavior_planner.cpp