feat: Save last path internally to enable smooth flexible transitions
The behavior planner now keeps an updated internal copy of the last planned path at all times. This allows for smooth transitions between the current path and a new path regardless of how many points of the current path should be taken over into the new path.

The current implementation works quite well, but it still shows suboptimal or illogical behavior in some situations and needs to be improved further. It also probably contains a few non-critical bugs that inhibit performance.